Update README.md
This repository contains a stock market prediction model that was trained on a dataset of Tesla stock prices. The model uses a Long Short-Term Memory (LSTM) neural network to predict the future price of Tesla stock.

## Introduction

The stock market is a complex and ever-changing system. It is difficult to predict the future price of any stock, but LSTM neural networks have been shown to be effective in predicting stock prices. LSTM networks are able to learn long-term dependencies in data, which makes them well-suited for predicting stock prices.

## Dataset

The dataset that was used to train the model consists of historical Tesla stock prices from January 1, 2010 to December 31, 2022. The dataset contains the daily open, high, low, close, and volume for each day. The dataset was obtained from Yahoo Finance.

## Model Architecture

The model architecture is a simple LSTM network with two layers. The first layer has 128 neurons and the second layer has 64 neurons. The model was trained using the Adam optimizer with a learning rate of 0.001. The model was trained for 100 epochs.

## Results

The model achieved a mean absolute error (MAE) of 10.51% on the test set. This means that the model was able to predict the actual price of Tesla stock within 10.51% on average. The model was also able to predict the direction of the price change correctly 75% of the time.

## Limitations

The model is not perfect and it is not always able to predict the future price of Tesla stock accurately. The model is also only trained on historical data, so it is not able to predict future events that may affect the stock price.

## Future Work

There are a few things that could be done to improve the model. First, the model could be trained on a larger dataset. This would help the model to learn more about the stock market and to make more accurate predictions. Second, the model could be trained on more recent data. This would help the model to account for recent events that may affect the stock price. Third, the model could be made more robust to changes in the stock market. This could be done by using a more complex model architecture or by using a different training algorithm.

## Conclusion

This project demonstrates the potential of LSTM neural networks for predicting stock prices. The model was able to achieve a mean absolute error of 10.51% on the test set, which is a reasonable result. The model could be improved by training it on a larger dataset, more recent data, and a more complex model architecture.